Refactor: Clean up Position and Velocity structs for better organization and clarity

+++ StellaEcs/AyanamisTower.StellaEcs.Components/Position2D.cs
@@ -6,11 +6,11 @@ namespace AyanamisTower.StellaEcs.Components;
 /// <summary>
 /// A basic wrapper around a vec2 type, exposed as Value
 /// </summary>
-public struct Position2D(int X = 0, int Y = 0)
+public struct Position2D(float X = 0, float Y = 0)
 {
     /// <summary>
     /// Access to the vector2 type
     /// </summary>
-    public Vector2 Value = new(X, Y);
+    public Vector2 Value  = new(X, Y);
 }
