+++ StellaEcs/AyanamisTower.StellaEcs.Components/Position3D.cs
@@ -5,7 +5,7 @@ namespace AyanamisTower.StellaEcs.Components;
 /// <summary>
 /// A basic wrapper around a vec3 type, exposed as Value
 /// </summary>
-public struct Position3D(int X = 0, int Y = 0, int Z = 0)
+public struct Position3D(float X = 0, float Y = 0, float Z = 0)
 {
     /// <summary>
     /// Access to the vector3 type
