+++ StellaEcs/AyanamisTower.StellaEcs.Components/Velocity2D.cs
@@ -6,7 +6,7 @@ namespace AyanamisTower.StellaEcs.Components;
 /// <summary>
 /// A basic wrapper around a vec2 type, exposed as Value
 /// </summary>
-public struct Velocity2D(int X = 0, int Y = 0)
+public struct Velocity2D(float X = 0, float Y = 0)
 {
     /// <summary>
     /// Access to the vector2 type
